signer: remove explicitly mentioned script type
The signing scripts will include the appropriate
hash-bang (#!) and executable permissions, allowing
them to be run directly. These permissions are
preserved when the scripts are copied into the
rpmbuild sandbox environment.

To support signing with the Gerrit
pipeline—which will use an ELF executable
version of the signer—this change is required.

Change-Id: Ic37b06dcc94c6c0bedd7d58298aa627a02ed0937
Signed-off-by: Kuntal Nayak <kuntal.nayak@broadcom.com>
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/26411
Reviewed-by: Michelle Wang <michelle.wang@broadcom.com>
Reviewed-by: Bo Gan <bo.gan@broadcom.com>
Tested-by: Michelle Wang <michelle.wang@broadcom.com>